+++ CHANGELOG.md
@@ -37,6 +37,8 @@ also non string parameters (e.g. `Enum.join([1, 2], ",")`
 - Support for directory listing using POSIX APIs: (`atomvm:posix_opendir/1`,
 `atomvm:posix_readdir/1`, `atomvm:posix_closedir/1`).
 - ESP32: add support for `esp_adc` ADC driver, with Erlang and Elixir examples
+- Add handler for ESP32 network driver STA mode `beacon_timeout` (event: 21), see issue
+[#1100](https://github.com/atomvm/AtomVM/issues/1100)
 
 ### Changed
 
@@ -55,6 +57,10 @@ instead
 - `unicode:characters_to_list`: fixed bogus out_of_memory error on some platforms such as ESP32
 - Fix crash in Elixir library when doing `inspect(:atom)`
 - General inspect() compliance with Elixir behavior (but there are still some minor differences)
+- Fix several uses of free on prevously released memory on ESP32, under certain error condition using
+`network:start/1`, that would lead to a hard crash of the VM.
+- Fix a bug in ESP32 network driver where the low level driver was not being stopped and resoureces were not freed
+when `network:stop/0` was used, see issue [#643](https://github.com/atomvm/AtomVM/issues/643)
 
 ## [0.6.4] - 2024-08-18